Init v2.0: - Remove event.GlobalClass and event.QueryClass - Add event.Class - Remove stat stddev - Rename stat Pct95 to P95 - Rename stat True to Sum - Rename event.Example to event.Sample - Rename event.EventAggregator to event.Aggregator - Change all float32 to float64 - Support outliers for Percona Server slow_query_log_always_write_time

+++ log/slow/parser.go
@@ -240,7 +240,7 @@ func (p *SlowLogParser) parseHeader(line string) {
 			if strings.HasSuffix(smv[1], "_time") || strings.HasSuffix(smv[1], "_wait") {
 				// microsecond value
 				val, _ := strconv.ParseFloat(smv[2], 32)
-				p.event.TimeMetrics[smv[1]] = float32(val)
+				p.event.TimeMetrics[smv[1]] = float64(val)
 			} else if smv[2] == "Yes" || smv[2] == "No" {
 				// boolean value
 				if smv[2] == "Yes" {
